Moving DerivedFnCollector out of ClangPlugin
