+++ lib/Differentiator/ReverseModeVisitor.cpp
@@ -1799,24 +1799,28 @@ Expr* getArraySizeExpr(const ArrayType* AT, ASTContext& context,
           if (DerivedCallOutputArgs[i + isaMethod])
             pullbackRequest.DVI.push_back(FD->getParamDecl(i));
 
-        FunctionDecl* pullbackFD = nullptr;
-        if (!m_ExternalSource) {
-          // Derive the declaration of the pullback function.
-          pullbackRequest.DeclarationOnly = true;
-          pullbackFD =
-              plugin::ProcessDiffRequest(m_CladPlugin, pullbackRequest);
+        FunctionDecl* pullbackFD =
+            m_Builder.FindDerivedFunction(pullbackRequest);
+        if (!pullbackFD) {
+          if (!m_ExternalSource) {
+            // Derive the declaration of the pullback function.
+            pullbackRequest.DeclarationOnly = true;
+            pullbackFD =
+                plugin::ProcessDiffRequest(m_CladPlugin, pullbackRequest);
 
-          // Add the request to derive the definition of the pullback function.
-          pullbackRequest.DeclarationOnly = false;
-          pullbackRequest.DerivedFDPrototype = pullbackFD;
-          plugin::AddRequestToSchedule(m_CladPlugin, pullbackRequest);
-        } else {
-          // FIXME: Error estimation currently uses singleton objects -
-          // m_ErrorEstHandler and m_EstModel, which is cleared after each
-          // error_estimate request. This requires the pullback to be derived at
-          // the same time to access the singleton objects.
-          pullbackFD =
-              plugin::ProcessDiffRequest(m_CladPlugin, pullbackRequest);
+            // Add the request to derive the definition of the pullback
+            // function.
+            pullbackRequest.DeclarationOnly = false;
+            pullbackRequest.DerivedFDPrototype = pullbackFD;
+            plugin::AddRequestToSchedule(m_CladPlugin, pullbackRequest);
+          } else {
+            // FIXME: Error estimation currently uses singleton objects -
+            // m_ErrorEstHandler and m_EstModel, which is cleared after each
+            // error_estimate request. This requires the pullback to be derived
+            // at the same time to access the singleton objects.
+            pullbackFD =
+                plugin::ProcessDiffRequest(m_CladPlugin, pullbackRequest);
+          }
         }
 
         // Clad failed to derive it.
@@ -1907,15 +1911,20 @@ Expr* getArraySizeExpr(const ArrayType* AT, ASTContext& context,
           clad::utils::ComputeEffectiveFnName(FD);
       calleeFnForwPassReq.VerboseDiags = true;
 
-      // Derive declaration of the the forward pass function.
-      calleeFnForwPassReq.DeclarationOnly = true;
       FunctionDecl* calleeFnForwPassFD =
-          plugin::ProcessDiffRequest(m_CladPlugin, calleeFnForwPassReq);
+          m_Builder.FindDerivedFunction(calleeFnForwPassReq);
+      if (!calleeFnForwPassFD) {
+        // Derive declaration of the the forward pass function.
+        calleeFnForwPassReq.DeclarationOnly = true;
+        calleeFnForwPassFD =
+            plugin::ProcessDiffRequest(m_CladPlugin, calleeFnForwPassReq);
 
-      // Add the request to derive the definition of the forward pass function.
-      calleeFnForwPassReq.DeclarationOnly = false;
-      calleeFnForwPassReq.DerivedFDPrototype = calleeFnForwPassFD;
-      plugin::AddRequestToSchedule(m_CladPlugin, calleeFnForwPassReq);
+        // Add the request to derive the definition of the forward pass
+        // function.
+        calleeFnForwPassReq.DeclarationOnly = false;
+        calleeFnForwPassReq.DerivedFDPrototype = calleeFnForwPassFD;
+        plugin::AddRequestToSchedule(m_CladPlugin, calleeFnForwPassReq);
+      }
 
       assert(calleeFnForwPassFD &&
              "Clad failed to generate callee function forward pass function");
