+++ tools/ClangPlugin.h
@@ -7,8 +7,8 @@
 #ifndef CLAD_CLANG_PLUGIN
 #define CLAD_CLANG_PLUGIN
 
-#include "DerivedFnInfo.h"
 #include "clad/Differentiator/DerivativeBuilder.h"
+#include "clad/Differentiator/DerivedFnCollector.h"
 #include "clad/Differentiator/DiffMode.h"
 #include "clad/Differentiator/DiffPlanner.h"
 #include "clad/Differentiator/Version.h"
@@ -37,41 +37,16 @@ namespace clang {
 
 namespace clad {
 
-  bool checkClangVersion();
-  /// This class is designed to store collection of `DerivedFnInfo` objects.
-  /// It's purpose is to avoid repeated generation of same derivatives by
-  /// making it possible to reuse previously computed derivatives.
-  class DerivedFnCollector {
-    using DerivedFns = llvm::SmallVector<DerivedFnInfo, 16>;
-    /// Mapping to efficiently find out information about all the derivatives of
-    /// a function.
-    llvm::DenseMap<const clang::FunctionDecl*, DerivedFns> m_DerivedFnInfoCollection;
+bool checkClangVersion();
+class CladTimerGroup {
+  llvm::TimerGroup m_Tg;
+  std::vector<std::unique_ptr<llvm::Timer>> m_Timers;
 
-  public:
-    /// Adds a derived function to the collection.
-    void Add(const DerivedFnInfo& DFI);
-
-    /// Finds a `DerivedFnInfo` object in the collection that satisfies the
-    /// given differentiation request.
-    DerivedFnInfo Find(const DiffRequest& request) const;
-
-    bool IsDerivative(const clang::FunctionDecl* FD) const;
-
-  private:
-    /// Returns true if the collection already contains a `DerivedFnInfo`
-    /// object that represents the same derivative object as the provided
-    /// argument `DFI`.
-    bool AlreadyExists(const DerivedFnInfo& DFI) const;
-  };
-  class CladTimerGroup {
-    llvm::TimerGroup m_Tg;
-    std::vector<std::unique_ptr<llvm::Timer>> m_Timers;
-
-  public:
-    CladTimerGroup();
-    void StartNewTimer(llvm::StringRef TimerName, llvm::StringRef TimerDesc);
-    void StopTimer();
-  };
+public:
+  CladTimerGroup();
+  void StartNewTimer(llvm::StringRef TimerName, llvm::StringRef TimerDesc);
+  void StopTimer();
+};
 
   namespace plugin {
     struct DifferentiationOptions {
